requirements.txt
# Discord Bot
discord.py>=2.0.0

# Stock Data API
yfinance>=0.2.0

# Utilities
python-dotenv>=1.0.0
aiohttp>=3.8.0
